refactor: split  shiftable module up into a calculatable module
